+++ feagi_connector_core/src/caching/mod.rs
@@ -1,9 +1,10 @@
 mod hashmap_helpers;
-mod io_cache;
-mod motor_channel_stream_cache;
-mod sensory_channel_stream_cache;
-mod io_cache_data;
+//mod io_cache;
+
+mod io_motor_cache;
+
 mod io_sensor_cache;
 mod per_channel_stream_caches;
+mod io_cache;
 
 pub use io_cache::IOCache;
